Modularize client directory structure
