--- client/src/app/components/NavBar.tsx
@@ -1,19 +1,0 @@
-import '@/app/styles/navbar.css'
-
-export default function NavBar() {
-    return (
-        <html><body>
-            <main>
-                <div className="topbar"> 
-                    <div className="nav-links"> 
-                        <a href="#">Dashboard</a> 
-                        <a href="#">Trades</a> 
-                        <a href="#">Watchlist</a> 
-                        <a href="#">Settings</a> 
-                        </div> 
-                        <a href="#" className="logout">Logout</a>
-                    </div>
-            </main>
-        </body></html>
-    );
-}
