+++ client/src/app/dashboard/layout.tsx
@@ -1,0 +1,15 @@
+import { ReactNode } from 'react';
+import NavBar from '@/components/NavBar';
+
+export default function DashboardLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <>
+      <NavBar />
+      {children}
+    </>
+  );
+}
